refactor: API 呼び出しの責務を ViewModel から分離 (#5)
