+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/MainFragment.kt
@@ -13,6 +13,7 @@ import androidx.lifecycle.repeatOnLifecycle
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.*
 import jp.co.yumemi.android.code_check.MainActivity.Companion.updateLastSearchDate
+import jp.co.yumemi.android.code_check.data.GitHubAPI
 import jp.co.yumemi.android.code_check.databinding.FragmentMainBinding
 import jp.co.yumemi.android.code_check.models.Repository
 import kotlinx.coroutines.launch
@@ -27,7 +28,7 @@ class MainFragment : Fragment(R.layout.fragment_main) {
 
         val binding = FragmentMainBinding.bind(view)
 
-        val viewModel = MainViewModel()
+        val viewModel = MainViewModel(GitHubAPI())
 
         // RecyclerView の登場人物を取得
         val layoutManager = LinearLayoutManager(requireContext())
